Use size_t or long instead of int, to avoid a loss of precision when these variables are being assigned values from similarly large integer types.  The cases where we used size_t have been checked to ensure the change of signedness should not cause problems (e.g., with negative values, or with signed/unsigned integer comparisons).

+++ source/libnormaliz/HilbertSeries.cpp
@@ -145,7 +145,7 @@ void poly_div(vector<Integer>& q, vector<Integer>& r, const vector<Integer>& a, 
     r = a;
     remove_zeros(r);
     size_t b_size = b.size();
-    int degdiff = r.size() - b_size;  // degree differenz
+    size_t degdiff = r.size() - b_size;  // degree differenz
     if (r.size() < b_size) {
         q = vector<Integer>();
     }
